Create Info area in player prompt
