+++ Components/GameBoard/GameBoard.jsx
@@ -472,6 +472,7 @@ export class GameBoard extends React.Component {
                             controls={ thisPlayer.controls }
                             promptText={ thisPlayer.menuTitle }
                             promptTitle={ thisPlayer.promptTitle }
+                            promptInfo={ thisPlayer.promptInfo }
                             onButtonClick={ this.onCommand }
                             onMouseOver={ this.onMouseOver }
                             onMouseOut={ this.onMouseOut }
