+++ less/gameboard.less
@@ -330,6 +330,11 @@
   margin: 0 5px;
   padding: 4px;
   text-align: center;
+
+  .alert {
+    margin-bottom: 0px;
+    padding: 2px 0 0 0;
+  }
 }
 
 .menu-pane {
